handwheel/linux/requirements.txt
cffi==1.16.0
gevent==23.9.1
greenlet==3.0.3
pycparser==2.21
pyserial==3.4
setuptools==69.0.3
websocket-client==1.6.3
websockets==11.0.3
zope.event==5.0
zope.interface==6.1
